+++ runtime/files/libuv/primop.c
@@ -25,9 +25,9 @@ yy_ptr yyIntDiv(yy_ptr i1, yy_ptr i2){
 yy_ptr yyIntToString(yy_ptr i1){
 
     int64_t num = addr_to_int(i1);
-    int64_t bufSize = snprintf( NULL, 0, "%lld", num ) + 1;
+    int64_t bufSize = snprintf( NULL, 0, "%lld", (long long )num ) + 1;
     char *strBuffer = GC_MALLOC(bufSize);
-    snprintf(strBuffer, bufSize, "%lld", num);
+    snprintf(strBuffer, bufSize, "%lld", (long long) num);
     return string_to_addr(strBuffer);
 }
